+++ NEWS.md
@@ -1,3 +1,21 @@
+## PPBDS.data 0.3.1
+
+---
+
+- Minor improvements to the `qscores` dataset. The 'course_name' and 'course_number' variables have been renamed to course 'name' and 'number'; whitespaces at the beginning of each course name have been removed; course terms are now in the format '2020-Spring' instead of '2020S'.
+
+
+
+
+## PPBDS.data 0.3.0
+
+---
+
+- Adds the `governors` dataset.
+
+
+
+
 ## PPBDS.data 0.2.0
 
 ---